# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/v1/models": {
      "get": {
        "summary": "获取模型列表",
        "deprecated": false,
        "description": "获取当前可用的模型列表。\n\n根据请求头自动识别返回格式：\n- 包含 `x-api-key` 和 `anthropic-version` 头时返回 Anthropic 格式\n- 包含 `x-goog-api-key` 头或 `key` 查询参数时返回 Gemini 格式\n- 其他情况返回 OpenAI 格式\n",
        "operationId": "listModels",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Anthropic API Key (用于 Claude 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "anthropic-version",
            "in": "header",
            "description": "Anthropic API 版本",
            "required": false,
            "example": "",
            "schema": {
              "type": "string",
              "example": "2023-06-01"
            }
          },
          {
            "name": "x-goog-api-key",
            "in": "header",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1beta/models": {
      "get": {
        "summary": "Gemini 格式获取",
        "deprecated": false,
        "description": "以 Gemini API 格式返回可用模型列表",
        "operationId": "listModelsGemini",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeminiModelsResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
	    "/v1/responses/compact": {
	      "post": {
	        "summary": "压缩对话 (OpenAI Responses API)",
	        "deprecated": false,
	        "description": "OpenAI Responses API，用于对长对话进行 compaction。",
	        "operationId": "compactResponse",
        "tags": [
          "OpenAI格式(Responses)"
        ],
        "parameters": [],
	        "requestBody": {
	          "content": {
	            "application/json": {
	              "schema": {
	                "$ref": "#/components/schemas/ResponsesCompactionRequest"
	              }
	            }
	          },
	          "required": true
	        },
        "responses": {
          "200": {
            "description": "成功压缩对话",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesCompactionResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/images/generations": {
      "post": {
        "summary": "生成图像(qwen-image)",
        "deprecated": false,
        "description": " 百炼qwen-image系列图片生成",
        "operationId": "createImage",
        "tags": [
          "图片生成/Qwen千问"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "input": {
                    "type": "object",
                    "properties": {
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "role": {
                              "type": "string"
                            },
                            "content": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "text": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "messages"
                    ]
                  },
                  "parameters": {
                    "type": "object",
                    "properties": {
                      "negative_prompt": {
                        "type": "string"
                      },
                      "prompt_extend": {
                        "type": "boolean"
                      },
                      "watermark": {
                        "type": "boolean"
                      },
                      "size": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              },
              "example": {
                "model": "qwen-image-plus",
                "input": {
                  "messages": [
                    {
                      "role": "user",
                      "content": [
                        {
                          "text": "一副典雅庄重的对联悬挂于厅堂之中，房间是个安静古典的中式布置，桌子上放着一些青花瓷，对联上左书“义本生知人机同道善思新”，右书“通云赋智乾坤启数高志远”， 横批“智启通义”，字体飘逸，在中间挂着一幅中国风的画作，内容是岳阳楼。"
                        }
                      ]
                    }
                  ]
                },
                "parameters": {
                  "negative_prompt": "",
                  "prompt_extend": true,
                  "watermark": false,
                  "size": "1328*1328"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功生成图像",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/images/edits": {
      "post": {
        "summary": "编辑图像(qwen-image-edit)",
        "deprecated": false,
        "description": " 百炼qwen-image系列图片生成",
        "operationId": "createImage",
        "tags": [
          "图片生成/Qwen千问"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "input": {
                    "type": "object",
                    "properties": {
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "role": {
                              "type": "string"
                            },
                            "content": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "image": {
                                    "type": "string"
                                  },
                                  "text": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "messages"
                    ]
                  },
                  "parameters": {
                    "type": "object",
                    "properties": {
                      "n": {
                        "type": "integer"
                      },
                      "negative_prompt": {
                        "type": "string"
                      },
                      "prompt_extend": {
                        "type": "boolean"
                      },
                      "watermark": {
                        "type": "boolean"
                      },
                      "size": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              },
              "example": "{\n    \"model\": \"qwen-image-edit-plus\",\n    \"input\": {\n        \"messages\": [\n            {\n                \"role\": \"user\",\n                \"content\": [\n                    {\n                        \"image\": \"https://help-static-aliyun-doc.aliyuncs.com/file-manage-files/zh-CN/20250925/fpakfo/image36.webp\"\n                    },\n                    {\n                        \"text\": \"生成一张符合深度图的图像，遵循以下描述：一辆红色的破旧的自行车停在一条泥泞的小路上，背景是茂密的原始森林\"\n                    }\n                ]\n            }\n        ]\n    },\n    \"parameters\": {\n        \"n\": 2,\n        \"negative_prompt\": \" \",\n        \"prompt_extend\": true,\n        \"watermark\": false\n    }"
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功生成图像",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/video/generations": {
      "post": {
        "summary": "创建视频生成任务",
        "deprecated": false,
        "description": "提交视频生成任务，支持文生视频和图生视频。\n\n返回任务 ID，可通过 GET 接口查询任务状态。\n",
        "operationId": "createVideoGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "kling-v1",
                "prompt": "宇航员在月球上漫步",
                "duration": 5,
                "width": 1280,
                "height": 720
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建视频生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/audio/speech": {
      "post": {
        "summary": "文本转语音",
        "deprecated": false,
        "description": "将文本转换为音频",
        "operationId": "createSpeech",
        "tags": [
          "OpenAI音频(Audio)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功生成音频",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/moderations": {
      "post": {
        "summary": "内容审核",
        "deprecated": false,
        "description": "检查文本内容是否违反使用政策",
        "operationId": "createModeration",
        "tags": [
          "Moderations"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功审核",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModelsResponse": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "example": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session认证，通过登录接口获取"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Access Token认证，格式: Bearer {access_token}，通过 /api/user/token 接口生成"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "用户ID请求头，必须与当前登录用户ID匹配，使用Session或AccessToken认证时必须提供"
      },
      "Combination": {
        "group": [
          {
            "id": "SessionAuth"
          },
          {
            "id": "NewApiUser"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          },
          {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}